Split yearFuncs into separate functions in year_funcs directory
